Added Functionality to Both Cone Stacks
Made the Cone Stack Selector Work
Took the functions for emission toggle off the robot objects and into each  individual object to prevent the toggle from being inaccurate.
Froze the cones in the stack after creation to prevent unknown falling.
Rethought the barriers for cone scoring and made it one barrier that just moves up rather than a new barrier clone everytime a cone scored.
Changed it so that the barriers are named based on their actual position on the field.